sdlayer: Place SDL_KEYDOWN events into keyasciififo[] when SDL_TEXTINPUT events are disabled.
This way, SDL_StopTextInput() does not break text entry for code that uses keyasciififo[] but does not currently call SDL_StartTextInput(), or if it cannot, such as Duke's cheat entry.

Technically we could revert r5326 and r5331, but since we only currently have support for showing ASCII text glyphs as bitmap fonts in ART tiles, there is no point to showing an IME on platforms that don't need an on-screen keyboard.

git-svn-id: https://svn.eduke32.com/eduke32@5968 1a8010ca-5511-0410-912e-c29ae57300e0