SW: Fix broken interpolation pathway following changes in 43ec16eb5576124bed3699943bb70b9ea943bd93.
What 43ec16eb5576124bed3699943bb70b9ea943bd93 intends to do was already done via 2d734664259e47b999fe989b924ecaa0a97bb109, however I applied the if statement in the opposite order so when it was cherry-picked again, the pathway was actually reversed - interpolating when alive and not when dead.

Rather than re-apply the opposite order, I've made the if statement match upstream for more consistency and to avoid further conflicts.