- added all needed parts of Witchaven's native looking menu.
This isn't particularly faithful, though, neither with DOS nor GDX because both menus have UX issues.

Main menu items reflect what a modern engine needs, not what DOS had.
The hi-res menu background is not being used because it forces the entire menu into the upper left corner.
The original game had the difficulty only as a setting on the "gore" page and could not start the game from there. This has been changed to a more traditional skill menu.
The only other part that has been ported over is the help screens.