- did some tests with the inside function and decided to only retain the 'old' one.
The one from EDuke32 is ambiguous - it considers any point on a wall to be part of both sectors touching that wall. This wasn't used anyway with the current engine compatibility settings.
I was not able to get different values out of the 'ps' version - this seems to only be important for demo playback concerns, for regular playback the differences appear to be totally irrelevant.

Ultimately this should be replaced anyway with a license-unencumbered variant of the same basic common algorithm.