SW: fix sector object setup.
When moving the SO out of the way to get the sector, this only works for rotatable SO's. For non-rotating ones the internally used functions never did what was expected, letting the SO end up in invalid space.
To do this properly the SOBJ_DONT_ROTATE flag must be cleared temporarily.