- Properly express all input constants mathematically so it's clear how they're derived.
Back in 2023, I knew all of these values like the back of my hand. Coming back in 2025, I had absolutely no idea what they were or how they were even calculated, especially the default push value for `scaletozero()` (which was also set up wrong along with `PITCH_HORIZOFFPUSH` as the ticrate was not factored in). This is all now corrected.